move StateModel to its own package
Move StateModel to its own package and rename AuthStateModel, with the
intent to make use of this code outside of just AuthProfile.
