+++ src/main/java/com/mozilla/secops/authstate/package-info.java
@@ -1,0 +1,2 @@
+/** Authentication state storage and utility classes */
+package com.mozilla.secops.authstate;
