+++ src/test/java/com/mozilla/secops/authstate/TestAuthStateModel.java
@@ -1,4 +1,4 @@
-package com.mozilla.secops.authprofile;
+package com.mozilla.secops.authstate;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
@@ -14,7 +14,7 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.contrib.java.lang.system.EnvironmentVariables;
 
-public class TestStateModel {
+public class TestAuthStateModel {
   @Rule public final EnvironmentVariables environmentVariables = new EnvironmentVariables();
 
   private void testEnv() {
@@ -24,26 +24,26 @@ public class TestStateModel {
     environmentVariables.set("DATASTORE_PROJECT_ID", "foxsec-pipeline");
   }
 
-  public TestStateModel() {}
+  public TestAuthStateModel() {}
 
   @Test
-  public void StateModelTest() throws Exception {
+  public void AuthStateModelTest() throws Exception {
     testEnv();
     State s = new State(new DatastoreStateInterface("authprofile", "teststatemodel"));
     StateCursor c;
     s.initialize();
 
     c = s.newCursor();
-    assertNull(StateModel.get("nonexist", c));
+    assertNull(AuthStateModel.get("nonexist", c));
     c.commit();
 
-    StateModel sm = new StateModel("riker");
+    AuthStateModel sm = new AuthStateModel("riker");
     assertNotNull(sm);
     c = s.newCursor();
     sm.set(c);
 
     c = s.newCursor();
-    sm = StateModel.get("riker", c);
+    sm = AuthStateModel.get("riker", c);
     assertNotNull(sm);
     assertEquals(sm.getEntries().size(), 0);
 
@@ -53,7 +53,7 @@ public class TestStateModel {
     sm.set(c);
 
     c = s.newCursor();
-    sm = StateModel.get("riker", c);
+    sm = AuthStateModel.get("riker", c);
     assertEquals(sm.getEntries().size(), 1);
     assertFalse(sm.updateEntry("127.0.0.1", 1.0, 1.0)); // Assert false for update existing
     sm.set(c);
@@ -63,11 +63,11 @@ public class TestStateModel {
     assertEquals(sm.getEntries().size(), 2);
     sm.set(c);
     c = s.newCursor();
-    sm = StateModel.get("riker", c);
+    sm = AuthStateModel.get("riker", c);
     assertEquals(sm.getEntries().size(), 2);
     c.commit();
 
-    sm = new StateModel("picard");
+    sm = new AuthStateModel("picard");
     assertNotNull(sm);
     assertTrue(sm.updateEntry("127.0.0.1", 1.0, 1.0));
     assertEquals(sm.getEntries().size(), 1);
@@ -75,17 +75,17 @@ public class TestStateModel {
     sm.set(c);
 
     c = s.newCursor();
-    sm = StateModel.get("picard", c);
+    sm = AuthStateModel.get("picard", c);
     assertTrue(sm.updateEntry("10.0.0.1", new DateTime().minusDays(1), 44.0, 44.0));
     sm.set(c);
 
     c = s.newCursor();
-    sm = StateModel.get("picard", c);
+    sm = AuthStateModel.get("picard", c);
     assertEquals(sm.getEntries().size(), 2);
     sm.pruneState(43200L);
     sm.set(c);
     c = s.newCursor();
-    sm = StateModel.get("picard", c);
+    sm = AuthStateModel.get("picard", c);
     assertEquals(sm.getEntries().size(), 1);
     c.commit();
